deleted unwanted sf page and reflected that in app
