+++ packages/app/src/App.tsx
@@ -15,11 +15,9 @@ import { I18nWrapper } from "@/components/I18nWrapper";
 import Intruder from "./pages/games/Intruder";
 import { IntroPage1 } from "./pages/StoryFactory/StoryFactoryPg1";
 import { IntroPage2 } from "./pages/StoryFactory/StoryFactoryPg2";
-import { IntroPage2a } from "./pages/StoryFactory/StoryFactoryPage2a";
-import { IntroPage3 } from "./pages/StoryFactory/StoryFactoryPg3";
+import { StoryFactoryPage3 } from "./pages/StoryFactory/StoryFactoryPg3";
 import { StoryFactoryPage4 } from "./pages/StoryFactory/StoryFactoryPg4";
 import { StoryFactoryPage5 } from "./pages/StoryFactory/StoryFactoryPg5";
-import { StoryFactoryPage6 } from "./pages/StoryFactory/StoryFactoryPg6";
 import Journeys from "./pages/Journeys";
 import { LanguageSwitcher } from "@/components/LanguageSwitcher";
 import Login from "./pages/Login";
@@ -251,7 +249,7 @@ const Router: React.FC = () => {
           path="/story-factory/2"
           render={() => (
             <AuthedLayout customBackground="#F7FAF9" wide={true}>
-              <IntroPage2a currentPage={2} />
+              <IntroPage2 currentPage={2} />
             </AuthedLayout>
           )}
         />
@@ -261,7 +259,7 @@ const Router: React.FC = () => {
           path="/story-factory/3"
           render={() => (
             <AuthedLayout customBackground="#F7FAF9" wide={true}>
-              <IntroPage3 currentPage={3} />
+              <StoryFactoryPage3 currentPage={3} />
             </AuthedLayout>
           )}
         />
@@ -271,7 +269,7 @@ const Router: React.FC = () => {
           path="/story-factory/4"
           render={() => (
             <AuthedLayout customBackground="#FBF2E2" wide={true}>
-              <StoryFactoryPage4 />
+              <StoryFactoryPage4 currentPage={4} />
             </AuthedLayout>
           )}
         />
@@ -286,7 +284,7 @@ const Router: React.FC = () => {
           )}
         />
 
-        <Route
+        {/* <Route
           exact
           path="/story-factory/6"
           render={() => (
@@ -294,17 +292,7 @@ const Router: React.FC = () => {
               <StoryFactoryPage6 />
             </AuthedLayout>
           )}
-        />
-
-        <Route
-          exact
-          path="/story-factory/6"
-          render={() => (
-            <AuthedLayout>
-              <StoryFactoryPage6 />
-            </AuthedLayout>
-          )}
-        />
+        /> */}
 
         <Route
           exact
