+++ packages/app/src/pages/StoryFactory/index.tsx
@@ -3,5 +3,3 @@ export * from "./StoryFactoryPg2";
 export * from "./StoryFactoryPg3";
 export * from "./StoryFactoryPg4";
 export * from "./StoryFactoryPg5";
-export * from "./StoryFactoryPg6";
-export * from "./StoryFactoryPage2a";
